Simplification of the matrix multiply pattern matching
Make use of the pattern acc_pattern = wild_f32x + f32(wild_f16x) * f32(wild_f16x) to match the whole update expresion

+++ test/performance/matrix_multiplication.cpp
@@ -37,7 +37,7 @@ int main(int argc, char **argv) {
     RDom k(0, matrix_size);
     RVar ki;
 
-    matrix_mul(x, y) += f32(A(k, y) * B(x, k));
+    matrix_mul(x, y) += f32(A(k, y)) * f32(B(x, k));
 
     Func out;
 
@@ -98,10 +98,10 @@ int main(int argc, char **argv) {
 
     double t = benchmark([&]() {
         out.realize(output);
-        
+
         if (target.has_gpu_feature()) {
             output.device_sync();
-        }        
+        }
     });
 
     // check results
